# src/main/java/cz/cvut/fit/tjv/fitnesscenter/model/SportType.java
package cz.cvut.fit.tjv.fitnesscenter.model;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.Id;
import jakarta.validation.constraints.NotBlank;
import lombok.*;

@Entity
@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@ToString
public class SportType {
    @Id
    @GeneratedValue
    public Long id;

    @NotBlank(message = "typeName is required.")
    private String typeName;
}